Added optional auto compile feature.
Added option auto compile feature which can be toggled in the File menu.

When auto compile is on the status icon moves down next to the lint
status icon, and the console will not automatically popup. Clicking the
status icon when auto compile is on will open the compiler log.

Also added manual compile button to the log window.